*: refactoring planner code #480 [summary] `planner` code has poor readability currently, we need refactoring the code. [test case] src/planner/builder/aggregate_plan_test.go src/planner/builder/builder_test.go src/planner/builder/expr_test.go src/planner/builder/from_test.go src/planner/builder/limit_plan_test.go src/planner/builder/orderby_plan_test.go src/planner/builder/sqlparser_test.go src/planner/others_plan_test.go src/planner/planner_test.go src/planner/select_plan_test.go src/planner/union_plan_test.go [patch codecov] src/planner/builder 97.7% src/planner 95.8%
