+++ src/executor/engine/operator/limit_operator.go
@@ -9,7 +9,7 @@
 package operator
 
 import (
-	"planner"
+	"planner/builder"
 	"xcontext"
 
 	"github.com/xelabs/go-mysqlstack/xlog"
@@ -22,11 +22,11 @@ var (
 // LimitOperator represents limit operator.
 type LimitOperator struct {
 	log  *xlog.Log
-	plan planner.Plan
+	plan builder.ChildPlan
 }
 
 // NewLimitOperator creates the new limit operator.
-func NewLimitOperator(log *xlog.Log, plan planner.Plan) *LimitOperator {
+func NewLimitOperator(log *xlog.Log, plan builder.ChildPlan) *LimitOperator {
 	return &LimitOperator{
 		log:  log,
 		plan: plan,
@@ -36,7 +36,7 @@ func NewLimitOperator(log *xlog.Log, plan planner.Plan) *LimitOperator {
 // Execute used to execute the operator.
 func (operator *LimitOperator) Execute(ctx *xcontext.ResultContext) error {
 	rs := ctx.Results
-	plan := operator.plan.(*planner.LimitPlan)
+	plan := operator.plan.(*builder.LimitPlan)
 	rs.Limit(plan.Offset, plan.Limit)
 	return nil
 }
