+++ src/planner/builder/from_test.go
@@ -6,11 +6,12 @@
  *
  */
 
-package planner
+package builder
 
 import (
-	"router"
 	"testing"
+
+	"router"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/xelabs/go-mysqlstack/sqlparser"
