+++ src/planner/builder/plan_node.go
@@ -6,7 +6,7 @@
  *
  */
 
-package planner
+package builder
 
 import (
 	"xcontext"
@@ -17,7 +17,7 @@ import (
 // PlanNode interface.
 type PlanNode interface {
 	buildQuery(tbInfos map[string]*tableInfo)
-	Children() *PlanTree
+	Children() []ChildPlan
 	getFields() []selectTuple
 	getReferTables() map[string]*tableInfo
 	GetQuery() []xcontext.QueryTuple
@@ -62,51 +62,4 @@ func findLCA(h, p1, p2 SelectNode) SelectNode {
 		return pr
 	}
 	return pl
-}
-
-// getOneTableInfo get a tableInfo.
-func getOneTableInfo(tbInfos map[string]*tableInfo) (string, *tableInfo) {
-	for tb, tbInfo := range tbInfos {
-		return tb, tbInfo
-	}
-	return "", nil
-}
-
-// procure requests for the specified column from the plan
-// and returns the join var name for it.
-func procure(tbInfos map[string]*tableInfo, col *sqlparser.ColName) string {
-	var joinVar string
-	field := col.Name.String()
-	table := col.Qualifier.Name.String()
-	tbInfo := tbInfos[table]
-	node := tbInfo.parent
-	jn := node.parent.(*JoinNode)
-
-	joinVar = col.Qualifier.Name.CompliantName() + "_" + col.Name.CompliantName()
-	if _, ok := jn.Vars[joinVar]; ok {
-		return joinVar
-	}
-
-	tuples := node.getFields()
-	index := -1
-	for i, tuple := range tuples {
-		if tuple.isCol {
-			if field == tuple.field && table == tuple.referTables[0] {
-				index = i
-				break
-			}
-		}
-	}
-	// key not in the select fields.
-	if index == -1 {
-		tuple := selectTuple{
-			expr:        &sqlparser.AliasedExpr{Expr: col},
-			field:       field,
-			referTables: []string{table},
-		}
-		index, _ = node.pushSelectExpr(tuple)
-	}
-
-	jn.Vars[joinVar] = index
-	return joinVar
 }
